requirements-gpu.txt
tensorflow-gpu==2.3.0
opencv-python==4.1.1.26
lxml
tqdm
argparse
easydict
matplotlib
pillow
ruanel.yaml